LPS-155715 SF, extract part of logic to BaseEhcachePortalCache, prepare for next
